removed unnecessary information from the CinemaHall class (createRooms method)
