+++ src/main/java/org/example/CinemaHall.java
@@ -1,6 +1,5 @@
 package org.example;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -16,21 +15,6 @@ public class CinemaHall {
 
     public String getTypeOfNumber() {
         return typeOfNumber;
-    }
-
-    public static List<CinemaHall> createRooms() {
-        List<CinemaHall> roomList = new ArrayList<>();
-        roomList.add(new CinemaHall("One", "2D"));
-        roomList.add(new CinemaHall("Two", "3D"));
-        roomList.add(new CinemaHall("Three", "4D"));
-        roomList.add(new CinemaHall("Four", "2D"));
-        roomList.add(new CinemaHall("Five", "3D"));
-        roomList.add(new CinemaHall("Six", "4D"));
-        roomList.add(new CinemaHall("Seven", "2D"));
-        roomList.add(new CinemaHall("Eight", "3D"));
-        roomList.add(new CinemaHall("Nine", "4D"));
-        roomList.add(new CinemaHall("Ten", "4D"));
-        return roomList;
     }
 
     public static Map<String, Long> getNumberOfRoomsOfEachType(List<CinemaHall> rooms) {
